back/requirements.txt
-i https://pypi.org/simple
flask==3.1.0; python_version >= '3.9'
flask-cors==5.0.0
gevent==24.11.1; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
numpy==2.1.3; python_version >= '3.10'
pulp==2.9.0; python_version >= '3.7'
python-decouple==3.8
python-dotenv==1.0.1; python_version >= '3.8'
scipy==1.14.1; python_version >= '3.10'
